+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 Mock HttpClient class with a unit test example
 
+## Fixed
+- `ConnectorSchemaDetailsResponse.Data.Schemas` type changed: 
+    - Old `map[string]ConnectorSchemaConfigSchemaResponse`
+    - New `map[string]*ConnectorSchemaConfigSchemaResponse`
+
 ## [0.6.3](https://github.com/fivetran/go-fivetran/compare/v0.6.2...v0.6.3) - 2022-07-20
 
 ## Fixed